Go from zellij back to tmux
Zellij is bloated and badly optimized compared to tmux (e.g. it does
frequent file writes. After a month of use the startup time for zellij
was 2 full seconds because its session cache took a long time to process.
I tried disabling session serialization, but zellij keeps doing it even
with the config setting telling it not to)